Don't error on anything inheriting from unittest.mock.Base
Summary:
We used to check for unittest.mock.Mock or unittest.mock.MagicMock, but turns out sticking with that doesn't cover all classes in unittest.mock. Check instead for inheritance from unittest.mock.Base.

NonCallableMock is special-cased because the mypy workaround for erroring against mocks was to stub this in typeshed to inherit from Any, when it really inherits from unittest.mock.Base.

Reviewed By: dkgi

Differential Revision: D7298076

fbshipit-source-id: 5695ee38890536b275c6dce6f514686e18c97e8f