only view validly structured values as annotations in signature instantiation
Summary:
Previously, whenever we saw a argument being passed where a type was being expected, we viewed it as a type.

i.e., if f has the signature
```
def f(typing.Type[_T])->_T:
  pass
```

when we call `f(zork)`, `zork` was assumed to be a class name that exists in the type order. This resulted in analysis crashes for the following
semi-common pattern:
```
classmethod
def construct(cls: typing.Type[_T}) -> _T:
  ...
```

a call to `Subclass.construct()` would lead us to add a `self_or_class` argument to the `construct()` call and
crash due to not resolving `self_or_class`.

For now, check if the parsed annotation is consistent with the type order via a newly added function `AnalysisTypeOrder.is_valid_annotation`.

The reason this new function exists rather than changing `AnalysisTypeOrder.contains` is that `contains` is used to check whether the type order
has a literal annotation, and get the index from there if so. The potential confusion is why the long comment, as well as the test is there.

Reviewed By: dkgi

Differential Revision: D7331355

fbshipit-source-id: 81f9a695c704593bfb22e1d950af14a8530b5229