don't let type variables escape function signatures
Summary:
Reduces analysis crashes to 1800.

We were crashing on analyses when trying to join type variables, which are not stored in the order,
with regular types. These originate due to not instantiating bottom when adding constraints to function signatures, necessary
for picking the right one. This change happens after the signature is picked, and essentially gets rid of type variables in the return
annotation and assumes that they are bottom. This lets us handle the following case:

```
a = set()
for item in a:
  ... # item should have type bottom here, because we don't know anything in it

Reviewed By: dkgi

Differential Revision: D7343201

fbshipit-source-id: 425b6e9ea533065041d97bb8d02a0bc60397c7a9