infer undefined function returns as Bottom
Summary: Reduces crashes to 1376. Viewing the value as Bottom rather than Object in all uninstantiated cases means that we suppress the typechecker in some cases - the alternative is crashing which means that we lose all errors for a function. I'm open to different ways of handling this (only make something Bottom if it's a type variable?)

Reviewed By: dkgi

Differential Revision: D7355698

fbshipit-source-id: 89d5e06b0ecabad34c7ab5b977bacdcd9d9e982b