introduce and use StatementVisitor
Summary:
Noticed that we had a lot of visitors in the environment of the form

```
let expression _ =
  ()
```

which indicates that all the expressions were traversed, even when unnecessary. Introduce a visitor
that only visits statements, and doesn't rely on Transformer to shave some time off the environment building.

I'm putting this diff out early as the idea came to me a bit late, and would like to get feedback before I start extensively testing this to make sure we're not visiting anything.

My last 3 runs on distillery with `make release`:
```
2018-03-21 22:50:54,820 PERFORMANCE Full environment built: 16.872918s
2018-03-21 22:51:27,909 PERFORMANCE Full environment built: 15.623879s
2018-03-21 22:51:55,857 PERFORMANCE Full environment built: 15.455379s
```

The last 3 runs with `make release` for the previous diff on the stack (untangling connect_type_order):
```
2018-03-21 22:54:23,734 PERFORMANCE Full environment built: 22.819453s
2018-03-21 22:55:00,557 PERFORMANCE Full environment built: 23.221585s
2018-03-21 22:55:45,896 PERFORMANCE Full environment built: 23.050327s
```

I think the increase in complexity is worth the performance improvements.

Reviewed By: dkgi

Differential Revision: D7363505

fbshipit-source-id: da85b29c50d53d15b6e03e66ee1dd326f57f5734