Fix class attribute error classification
Summary: Previous addition of this detail assumed that class_variable field of Attribute was correctly set to true/false depending on the access pattern... turns out, it always defaults to true when the attribute is not defined. We're aware when this is sometimes false (depending on whether the access chan up to the attribute is a meta type); set the class_variable attribute according to this instead.

Reviewed By: sinancepel

Differential Revision: D7381845

fbshipit-source-id: b3ae4ab326b79ce4d6ad86f20e2864cdc8b84fa1