initial doc skeleton
Summary:
Description:

Initial cut of a 'Installing PyreCheck' page within a
Docusaurus skeleton.

Note that there's a .gitignore file here; it looks liked
the default that comes with Docusaurus is a little... off?

I'll fix that up separately once I figure that out, and also so that I
can rerun `generate_hgignore.py` in peace...

Reviewed By: shannonzhu, dark

Differential Revision: D7335656

fbshipit-source-id: f15a66bebd612aae8315e12c3a6bb71078bc127b