`remove_batch` before adding in the `register_` functions.
Summary:
Noticed while investigating a difference in reported errors when `PYRE_USE_SHARED_MEMORY=1`. The
`register_` functions always need to remove before adding. A longer term solution is to add a `set` function to the hack parallel functors,
but this is a bandaid that helps me debug my immediate issue.

Reviewed By: shannonzhu

Differential Revision: D7406818

fbshipit-source-id: fbca7df6c2e89528581509e30d75ac7f66cc4e7c