more exception handling on request handler thread
Summary:
We were not handling exceptions on the request handler thread. The result was that when it raised an exception, OCaml didn't
catch it at all, and kept chugging along with the other thread after printing the thread's backtrace.

Catch *all* exceptions here in order to make sure that both threads stop cleanly. We are not being more precise to be future-proof - catching
an exception here is a last resort.

Reviewed By: dkgi

Differential Revision: D7418158

fbshipit-source-id: fb33124fba658a51cd2c03d451c4d3e72836732e