Refactor ast.Attribute and attribute-collecting Class methods
Summary:
Don't represent attributes as just an assign + async. We want to know more information about it, and don't need everything in the assign. Structure already in place for this, and now we don't have to hackily create assigns out of thin air to  stand in for attributes that weren't defined with assigns anyway.

Will add a map in a diff on top of this to combine setters with their getter counterparts.

Reviewed By: dkgi

Differential Revision: D7422516

fbshipit-source-id: 42eeae1de2e59d129064bd15126958f481cbf812