Expand Transformer features
Summary:
The Transformer module allows simple manipulation of the AST.

The algorithm is roughly the following, starting from the top node:

Step 1) When you reach (=visit) a node X, recurse into all of its
        children first and visit them.

Step 2) Once the children of node X have been visited fully, call the
        expression or statement method on the node X itself, by
        providing the updated X as a parameter. Please note the
        children of X have already been visited and possibly modified
        because of step (1).

Step 3) Return to the parent of X the (possibly modified) node that
        was returned by step (2).

Two features are missing in this model:

* Be able to call a method on X before its children have been
  modified. Right now the only available hook point (expression or
  statement) is provided with a version of X whose children have
  already been irrevocably modified. There is no way to inspect the
  state of the children of X before they were transformed.

* There is no way to halt the recursion mid-way. In other words, the
  AST is always traversed fully, and there is no way to stop the
  recursion under a given node X if, for any reason, this recursion is
  not needed or (worse) undesired.

The solution implemented in this diff provides both features.

This is achieved by having an additional ..._preorder hook for
statements. This hook is executed on a node X before the children of X
are visited. Consequently, the existing expression and statement hooks
are renamed to expression_postorder and statement_postorder to clarify
their preexisting meaning. Note also that, similarly to what the
existing hook does, ..._preorder is allowed to modify node X or any of
its children, so ..._preorder must return Statement.t. This is not
implemented for Expression.t because it is not needed yet.

A statement_keep_recursing hook is added, that controls the recursion
listed as Step (1) in the current algorithm.

Reviewed By: dkgi

Differential Revision: D7347677

fbshipit-source-id: b9e40369e734a658909859a543528bc8e076c94b