Fix handling of 'autogenerated' config item
Summary:
Don't just blindly overwrite autogenerated with an empty list, but
read the config item only if a preceeding file did not set it already.

As part of testing, fix `test_init`, that was actually not testing
things as expected - need to mock `open()` so it always succeeds and
the two side effects of `json_load` actually matter.

Reviewed By: dkgi

Differential Revision: D7474837

fbshipit-source-id: fe059468c166925dcce37e720356886af9c65bed