encode implicit arguments in callable
Summary:
the rules are pretty simple: if the callable is called on an object
then it takes that object as an implicit parameter. Except that there is one
exception to this rule that needs to be dealt with :(

I considered dropping the implicit argument or adding a dummy argument but that
breaks down with reassignments of functions.

Reviewed By: sinancepel

Differential Revision: D7470542

fbshipit-source-id: 351aa4e33e9d97b5a370350a571b9b34064ef8c0