Set local annotation to type of attribute value
Summary:
Problem: we were getting unknowns where we don't expect in attributes: The code in P59345948 has a got unknown error instead of got int error.

I believe this accounts for many of the random "extraneous ignores" I was seeing in one_world when trying to upgrade.

Turns out we are throwing away the value field when creating attributes out of assigns. The code above works if int is annotated, but not if an integer value is assigned. I don't know if there was a reason for this, but it seems incorrect since we construct attribute annotations as global-annotation = type annotation, local-annotation = type of value assigned.

Reviewed By: sinancepel

Differential Revision: D7487902

fbshipit-source-id: 3ce162c724134f56b59ff115d5729ddb8bc4a1c6