Fix UndefinedType false positives in methods of union types
Summary:
We were throwing incorrect UndefinedType/Function errors when accessing methods on Union types. Tests and task have examples.

Still some work to be done to catch cases where a method (or attribute, which has the same temporary patch as this one: D7448417) is not defined on some types in the Union but maybe not others. Task: T27165573

Reviewed By: sinancepel

Differential Revision: D7523216

fbshipit-source-id: 78e40eb6541d4b878f6a11614fd8b30e8a6b661e