only cleanup edges for annotations that have been newly added
Summary:
We were iterating through all the vertices in the type order when connecting annotations
to top and removing extra edges, even though what we really care about is doing this postprocessing for the newly added edges. Keeping track of what's been added when registering class definitions and
only running the postprocessing on those reduces the typechecking time for an empty file with no dependencies from 0.30s to 0.07-0.06s.

Reviewed By: dkgi

Differential Revision: D7527046

fbshipit-source-id: eb37c6c0b3d0b589f732dd92f0b34fc904c31fa8