fix inference of constraints for callables
Summary:
We were converting types to classes and using the class constraints to get the set of generics, and zip them up with the parameters, assuming equal length. This isn't great for the following reason:

We'll miss cases where you have a subtype with a different set of type variables. Consider `d.values() : ValuesView[str, int} <: Iterable[int]`, used to instantiate a callable of type `[[Iterable[_T}, _T]`. With the previous implementation where we take a look at the generics of the class, we just collect the generic `_T` and try to hammer in the parameters. What really needs to be done here is to walk the type order up from the subclass to the superclass and instantiate accordingly, and then check if any type variable constraints are being violated.

Instantiating the parameters from `ValuesView[str, int] -> Iterable` will give us a type parameter instantiated to `int`, which is what we want.

Reviewed By: dkgi

Differential Revision: D7524108

fbshipit-source-id: dac782e0557e513b58ae163107a1215fe3adc95f