try to kill a server if it fails to stop
Summary:
One common issue with `pyre restart` is that a stop would fail, and you'd get a series
of `Client exited with status 1` exceptions. Attempt to improve this situation by trying a kill in the stop command
if the nice way of stopping fails. The justification is that when a user "stops" a server, they don't
care whether we kill the pid or stop via the request, and would just like the old server to go away.

In order to accomplish this, do 2 things:

1. Make the stop command in the binary return a non-zero error code on failure.

2. Handle client exceptions in the stop command and trigger a kill when that happens.

Reviewed By: dkgi

Differential Revision: D7890015

fbshipit-source-id: 3f2a3f4c377f4d96d33bd804371fec7c05752bc8