make the watchman client stop more consistently when the server dies
Summary:
The watchman client wasn't picking up on the fact that the server had died because
it was never listening on the server socket, and all that happened is that writes to the server socket
silently failed. (Yay, unix sockets). In order to solve this issue, make the watchman client read
the server socket as well, and read/throw away server values (the specification's that you need to read EOF).

Also handle EPIPE better as a consequence. We don't want to crash via SIGPIPE, we'd
rather stop the watchman service explicitly.

Reviewed By: dkgi

Differential Revision: D7884940

fbshipit-source-id: 4525f945e2d9a38c11b65807157c8f9167af3a18