Move tracking of number of workers into configuration
Summary:
The previous code was storing the number of workers into the arguments
data structure, and adding this information at runtime. This would
cause scary backtraces if the code raised an exception before the
default configuration was applied.

This diff moves the attribute into the configuration class, where it
is easier to track its default and its lifecycle. This way, backtraces
are not printed anymore.

Reviewed By: shannonzhu

Differential Revision: D7921335

fbshipit-source-id: ec1578ae0e42dac334def34f641ede4caea8077e