simplify vscode plugin
Summary:
Our previous vscode plugin, adapted from the vscode example by microsoft, was unpublishable as there
were some environment quirks that didn't let us only pull in the client and start it by default. Remove the old one and use the hack-vscode package as a template instead for the insipration.

Reviewed By: dkgi

Differential Revision: D7944096

fbshipit-source-id: 86a199cf15aef20b0ee4542cb294aad66ed62fd5