don't try to infer globals
Summary: We are not consistent in the way we order files across machines, so we could get inconsistencies due to resolving globals from a different module on one machine, and being unable to resolve in another, just based on which module we registered first. I have a task T29185210 for it.

Reviewed By: dkgi

Differential Revision: D7946759

fbshipit-source-id: 32fed041a9650c1f6af99cbbaa000970b2f9a933