Add support to build wheels for older versions of MacOS
Summary:
MacOS wheels used to support only OSX 10.13. This diff adds support to
compile the toolchain and binary with the MACOSX_DEPLOYMENT_TARGET
flag set to 10.11, thereby allowing binaries to run on older Macs.

This refers to #11, but it does not solve it yet. That issue will be
closed once we release the new version of the package.

Reviewed By: dkgi

Differential Revision: D7986807

fbshipit-source-id: 63633f5107a11a879bdc2f92aa4187f673870172