Instantiate type variables in parametrics when joining with primitive
Summary:
Fixes #40

This issue was revealed in the following case:

```
def foo(input: str) -> typing.List[int]:
    return [a for a in input]
```

Which would type error as "got List[Any], expected List[int]" instead of the expected error "got List[str], expected List[int]".

Turns out this is because when we resolve comprehensions, we join the target being iterated over (in this case, `input`) with `typing.Iterable[_T]` - which makes sense to resolve subtypes of iterables like `str` up `typing.Iterable[str]`.

Join logic would just check which type was leq the other, and return the greater type. But the leq logic instantiates typevariable params of the supertype using the subtype - in this case to get Iterable[str], which is correct -- but compares to the original of Iterable[_T] and decides that str is not leq Iterable[_T]. This logic works in most cases but isn't expecting uninstantiated type variables to be passed into the comparison.

Solution is to instantiate any Bottom type params of the parametric type before we check leq. This wasn't a problem when `input` is a parametric already b/c this instantiation happens in join of two parametrics.

Note: changes to testing env are based on typeshed builtins.pyi and typing.pyi - we were incorrectly inheriting str from Iterator instead of Iterable and bypassed some steps in the middle.

Reviewed By: dkgi

Differential Revision: D8030337

fbshipit-source-id: d6f31ff35ddcdd9322183805023b54cf7c996ba9