Honor TERM variable signaling non-capable terminals
Summary:
The Pyre client used to emit escape sequences regardless of the
terminal capabilities. The only workaround to avoid this behaviour was
to use `--noninteractive` and/or `--debug`.

This diff adds support in the client to recognize "dumb"
terminals. Also, it removes the implicit dependency between
`--noninteractive` and `logging.DEBUG`, so full debugging output is
emitted only if requested.

This fixes #45.

Reviewed By: dkgi

Differential Revision: D8017045

fbshipit-source-id: 57fc180a46f7a44a1fe43c4862b450fe59c08f9b