Do not print stray WRAP_OVERFLOW sequence before errors listing
Summary:
The log.cleanup() function was incorrectly printing a WRAP_OVERFLOW
escape sequence just before the final output. This was both
unnecessary (since TimedStreamHandler.terminate() does it) and
disruptive when the output was redirected to a file, rather than
visualized in a capable terminal.

Remove the unnecessary code, and move the call to flush() right where
it's needed.

Reviewed By: sinancepel

Differential Revision: D8022160

fbshipit-source-id: 1208dc54aad52a3b2b516aaf2fd3b783ce29381e