Disable progress reports on dumb terminals
Summary:
A directory with plenty of source files might generate a long list of
progress updates. When looking at that output in a dumb terminal, or
when stderr is not a TTY, that's a lot of clutter for nothing.

This diff changes the progress reporting log to use a new `Progress`
section, and adds code in the client to suppress these messages when
the "capable terminal" check fails.

Reviewed By: sinancepel

Differential Revision: D8021889

fbshipit-source-id: 2ea4d03e5de14f3391a8f490f0a58660dfc5a807