Display parse errors via commandline switch
Summary:
A file with parse errors cannot be type-checked. It is therefore
critical to always display the number of parse errors occurred.

Furthermore, it's useful to display those parse errors if the user
needs them. Hence, add a new commandline switch to support this
workflow.

A separate commandline switch at the client level is more effective
than asking the user to fiddle with `-logging-sections`, as the latter
is less descriptive in this case. That said, the user can still use
`-logging-sections parser` to achieve the same result.

Reviewed By: dkgi

Differential Revision: D8098637

fbshipit-source-id: 17286a2494007d6bb3275cf27e539cca1f3964b1