keep path information for modules around
Summary:
We were doing a janky "replace . with /, add a .py and check" algorithm for finding the path
corresponding to a module when computing dependencies. Let's store this information when constructing a module
so that a qualifier gives us a canonical path to look to. A upcoming diff will change the way we look up
dependency paths.

Reviewed By: dkgi

Differential Revision: D8083747

fbshipit-source-id: df805e0132f8fa8c7fdc15c6312d086c62b9aabc