only re-parse files that change environment in server
Summary:
We had a bug in server request handling that meant that only files getting re-checked were getting re-parsed. If you happened to add
a file which is supposed to update the environment but not check (such as a stub) we'd never reparse this file. On the contrary, we
would re-parse files that are probably unchanged (if they are changed, either watchman or a didSave notification will warn us in the future).

This didn't bite us beforehand because for watchman requests, we had always set the update_environment_with files = check files, and you only need to
re-parse one to add to shared memory.

Reviewed By: dkgi

Differential Revision: D8086320

fbshipit-source-id: 7a0ab6b026b7a8511e533c2bc24b80a67c24edad