update pyre check to return nonzero exit code when errors are found
Summary:
I didn't see an obvious way of using `pyre check` with scripts or a CI system. This change updates `pyre check` to return a nonzero exit code if it sees any errors. This exit code can then be used to fail the build if errors are seen.
Closes https://github.com/facebook/pyre-check/pull/9

Reviewed By: dkgi

Differential Revision: D7982524

Pulled By: dark

fbshipit-source-id: f8d00dab3b98b50e397203ee8d6c569dbe68a086