change incremental behavior to not restart server
Summary:
Justification: When the server crashes, Nuclide restarts it automatically, and the incremental mode restarting it as well would
trample the previous effort. The difference between typing "Y" and enter vs. re-running the command is negligible from the user side,
and this nudge would cause subsequent "incremental" calls to pick up on the running nuclide server.

Reviewed By: dkgi

Differential Revision: D8206509

fbshipit-source-id: be42087e802e6b5a9bf51546320c5a9d916cbb62