Add `-typeshed` commandline option
Summary:
So far, we were adding the `stdlib` directory to the search path,
expecting it to include directories like `2`, `3`, `3.x`, etc... Thus,
it was necessary to manually add `third_party` to the search path to
get its contents.

This diff adds a commandline option for the binary to receive the
typeshed root directory. The typeshed directory is expected to include
a series of directories (i.e. `stdlib`, `third_party`), that will be
scanned as part of the search path. This gives greater flexibility in
adding more directories to the search path (e.g., the aforementioned
`third_party`), by picking them directly from typeshed. If in the
future we will need to package more stubs, we will simply add them to
the typeshed directory.

Obligatory ASCII art to explain the directory structure:

```
  new "-typeshed" option
        points here         old "-search-path" option
                 |      +-- points here
                 |      |
                 V      V
/path/to/typeshed/stdlib/2/
                 |      /3/
                 |      /3.3/
                 |      ...
                 |
                 /third_party/2/
                             /3/
                             /3.3/
                             ...
```

Reviewed By: sinancepel

Differential Revision: D8216818

fbshipit-source-id: dc041bd4657060bd6336743c67a182821771811f