universal qualification
Summary:
This is a more or less complete rewrite of the qualification logic. Doing a single pass, no longer relying on the visitor module to give us better control over what's happening here.

The goal of this is to no longer special case how we qualify the top-level function. We're currently not type-checking that by default because the qualification of it is too broken.

Reviewed By: sinancepel

Differential Revision: D8068096

fbshipit-source-id: 99ce4719f6d96e27895064f399c3b09c757b49db