Track original directory and source root separately
Summary: If you call pyre on a target while cwd is fbcode, it seems fine that the local source root is considered fbcode - all errors in fbcode will be shown. But if you pass in a local config while cwd is fbcode, the local source root should be considered the path to that local config.

Reviewed By: sinancepel

Differential Revision: D8299731

fbshipit-source-id: 7b5bae8b45df1833be80e172c10b5031069f6736