don't use target for variables when registering
Summary:
This used to be necessary when we weren't qualifying strings. E.g. we would
qualify to `$local_0_T = typing.TypeVar('T')`. The right hand side of this
would parse to `Variable T` (not qualified) and we'd have to replace the name
with the qualified name of the target (during registration `qualifier.T`).

Now we get the right string and thus the right variable without extra magic.

Reviewed By: shannonzhu

Differential Revision: D8409607

fbshipit-source-id: 81cb66355856449bb0d8b69e283f9cd26f729605