Do not hash locations in Node.t
Summary:
`Node.compare` and `Node.equal` ignore the `location` field of the `t`
record, but the autogenerated `Node.hash` does not. This would cause
an inconsistency where `Node 1 == Node 2`, but `hash(Node 1) !=
hash(Node 2)`.

This diff fixes the inconsistency and adds tests.

This diff also fixes the inclusion of the `Hashable` module in
`AstExpression.Access`, that was using the wrong hash function. This
is tested too.

Reviewed By: dkgi

Differential Revision: D8504669

fbshipit-source-id: 1819aa2e827b212e795e808f6c996bc7a66a0849