Clear previous type resolution info upon recheck
Summary:
When the check of a file is retriggered in the server, clear all type
resolution info from shared memory for all affected sources. This
allows the workers to update the shared memory with the new
information upon recheck.

I believe this is an incomplete solution, because it removes accesses
from AnalysisTypeResolutionSharedMemory based on the *current* content
of the file, hence it will not remove definitions that have been
deleted from the file since the last save.

This is better than nothing for now (and I'd like to talk with rvantonder
about possible solutions first, before changing things further).

Reviewed By: sinancepel

Differential Revision: D8650026

fbshipit-source-id: 9b9824a491a0d96b1dded517acdeae4c42a65896