Back out "Refactor the parser to store location information in Access"
Summary: This is a short term concession we're making for the current upgrade - I'm working on a refactor that's going to use ints for paths in locations rather than strings, and minimize the perf effect of the refactor. (Right now, there's a penalty around 45-50 seconds). Will re-introduce this diff after the upgrade.

Reviewed By: dkgi

Differential Revision: D8672478

fbshipit-source-id: a5cea361430ad823f2eaaba2468f336928ba84e8