Write through the cache when batch-adding to shared mem
Summary:
Shared memory tables use two levels of caches to speed up accesses.
The first cache is an LRU cache capped at 1000 entries; the second has
a nominal capacity of 1000, and uses a LFU eviction strategy when size
reaches double that capacity.

When batch-adding to the tables a number of entries that exceeds the
capacity by several orders of magnitude, a lot of time would be spent
adding and evicting from the caches. Furthermore, since the table add
is performed in essentially random order, both caches would end up
with a random entry set, that does not benefit from either cache
strategy.

Change the code in serviceEnvironment.ml to write through caches
instead. Caches will be populated during access instead, where the
usage pattern is more evident and the benefit of LRU/LFU caches is
more pronounced.

Reviewed By: sinancepel

Differential Revision: D8699716

fbshipit-source-id: 57d292831a144e6f8fbb61988c8a3b75cd614000