Build lookup table ondemand (uncached)
Summary:
In the server process, ignore stored lookup information and build a
lookup table ondemand, based on the content of the hover request.

The final goal is to have these tables cached for all open files, but
I am publishing this diff early to help with removing lookup
dependencies in the type checking code and reduce its runtime.

Reviewed By: sinancepel

Differential Revision: D8722955

fbshipit-source-id: 154b6734d585c132c161575530bc39d94f1e336c