Remove default check current directory behavior when no sources specified
Summary: Stop warning when calling pyre from subproject with no local configuration (just write to debug output and go look for the .pyre_config in project root). Get rid of need for nonexistent directories in project root to default to noop, and also get rid of default to checking "." when nothing is provided to analyze. Prompt user to create a local config with `pyre init` or pass something in via command line instead.

Reviewed By: dark

Differential Revision: D8736266

fbshipit-source-id: 4fedb99fa416692e295a974ba9fc211c69aeeb70