generalize qualifier normalization for imports in modules
Summary:
We were handling a limited set of relative imports for modules - namely, special-casing
"from ." and "from builtins". Generalize this by supporting arbitrary dots to indicate moving up directories.

The function to normalize imports is pulled out because I intend to use this in Analysis.Preprocessing later. Can inline for
now and pull it out in a subsequent diff if desired.

Reviewed By: dkgi

Differential Revision: D8811286

fbshipit-source-id: f074043bca926dd21433810bcdde99cfba55565f