Fix purging of class parents in shared memory
Summary:
The server was not picking up updates to the type hierarchy of
classes. Predecessors/successors relations were added to the
TypeOrder, but they were never removed. Hence, the removal of a parent
definition for a class would never be reflected in the TypeOrder.

Of course this would work correctly for a one-shot `pyre check`.

Fix the issue by clearing the successor dependencies for all classes
that are part of a file that is reparsed by the server.

Predecessors cannot be easily cleaned, without tracking which files
contain classes that need reparsing; a newly added test demonstrates
the issue.

Reviewed By: sinancepel

Differential Revision: D8796619

fbshipit-source-id: e7afd16d6661f975ac8baba5e0c0bccc59486d19