be more precise in the pid we wait for
Summary: This is part 2) of triaging server crashes, where we tackle the issue where the server was waiting on pids it shouldn't have waited for. We used to have a SIGCHLD handler which would wait for all pids, but this also accidentally reaped worker processes and crashed pyre. Instead, store the pid of the at most 1 watchman client that pyre spawns in connections, and reap it when we get an End_of_file from a socket.

Reviewed By: rvantonder

Differential Revision: D8871851

fbshipit-source-id: 9db6dd548eaefb77949dd588c0be7bc1c09f1c49