Fix off-by-one column counting in the LSP server
Summary:
D7608655 incorrectly assumed that Pyre starts counting row:column at
1:1. In the internal representation, the first character of a file is
1:0 instead.

Hence, fix the math in `languageServerProtocolRequestParser.ml` that
would incorrectly convert from the LSP wire range protocol (that
starts at 0:0) into a protocol that starts at 1:1.

Fix also an off-by-one comparison in `analysisLookup.ml`, that was
masking the bug above. The semantics of the `stop` field of the
`Location.t` record are that it should point at the first character
_after_ the selected range. Hence, for a location to contain a
position, the stop column of the location must be strictly greater
than the column of the position.

Reviewed By: sinancepel

Differential Revision: D8888560

fbshipit-source-id: 06e9d823e2132fdcfaf0aa06804c27a9b5abf08f