start instead of restart in persistent client, and wait for previous starts to finish
Summary:
Prevents a few race conditions.

a) A restart() from the persistent client may be clobbering a restart from the command line, which would be a common reason for the server
socket to close. Mitigate this by calling start().

b) We had a weird case of "Warning: Server exists" followed by a "No server found" (also appears in incremental). One cause of this is due to having a "pyre start" hold on to the client lock before it
has a chance to start a server, and the start command therefore failing silently.

Mitigate this by polling for the client lock being free. Note that this doesn't rule out this category of race, but is a heuristic. In the worst case, the start fails, the call to the persistent fails, and Nuclide re-runs "pyre persistent".

Note that this requires us to switch the client_lock acquisitions to be blocking. Otherwise, we'd be introducing a race where persistent polls for the client lock, the start() command fails to acquire the lock because of this and fails.

Reviewed By: dark

Differential Revision: D8897525

fbshipit-source-id: 61e85fe94c483eb1f8e96af8abcc045dc42d480b