Warn on unrecognized configuration items
Summary:
If a configuration item is misspelled in a configuration file, its
absence might go unnoticed. For example, typing "target" rather than
"targets" will generate a cryptic error message (it happened to me
yesterday).

Add code so that, when a configuration file contains entries that are
not recognized, an appropriate warning is printed.

Reviewed By: shannonzhu

Differential Revision: D8942551

fbshipit-source-id: dd3c5a29abb19486c6e19c090d3170fbd581ac26