move filter_directories to ServiceTypeCheck
Summary:
One issue with the previous filter_directories implementation was that we would still pollute coverage data
for ignored files. Filtering the analyzed files (after an environment has been built) means that
the coverage data is more accurate, as well as saving time and allowing our editor integration to be more accurate.

Reviewed By: dkgi

Differential Revision: D8940850

fbshipit-source-id: 2c7fcff81634c9eb1032d26f656a2ef5e1a5bb7e