Introduce explicit LSP requests for opening, saving, closing a file
Summary:
The rationale behind this change is to be able to perform more actions
in the server upon the receipt of these requests: for example, caching
lookup tables and source code, rather than re-creating them for each
request.

Furthermore, rather than letting the parser take decisions on what to
do with a request (e.g. conditionally handle a SaveDocument based on
the `check_on_save` flag), move that logic into the server.

Reviewed By: sinancepel

Differential Revision: D8883109

fbshipit-source-id: efc64dec5026677ea804896abd65df21676715bb