infer files to analyze based on qualifier instead of path
Summary:
Previously, we were conjuring a random "qualifier.py" path whenever `module.py` had an `import qualifier` statement.
This is a problem if `qualifier.py` didn't exist beforehand or was a `__init__.py` file, as we'd look for the wrong file.

This diff stores the import qualifiers directly, and whenever `module.py` imports `qualifier`, adds a `Access qualifier -> [module.py; others]` mapping.

When a file is updated in the server, it computes the qualifier for the path and looks this qualifier up.

Reviewed By: dkgi

Differential Revision: D9008868

fbshipit-source-id: 8bf97de652bf69c7df4248404930b588ae4dee59