Fix implicit attribute tuple assignment bug
Summary:
facedown_cat

We were unpacking tuples incorrectly...

We used to just treat tuples as a list of single assignments, and register the attributes as per usual. The problem is, we also have an optimization (added per request from IG developers) that shortcuts annotating attributes of the format:

```
def foo(self, x: int):
  self.something = x
```

By just grabbing the annotation from the parameter and adding it to the attribute assignment.

This meant that:

```
def foo(self, x: Tuple[int, int]):
   self.a, self.b = x
```

This would create attributes a and b, which both have value x and are annotated to each be `Tuple[int, int]`.
** see attached task for example from wild

This change simply does not do that parameter annotation optimization in the case that we are unpacking tuples. Originally, I wanted to edit the value as well to add on a `.__getitem__[index]`, but that's not very clean because we have unhandled cases such as `x, *y, z = range(10)`.

Additionally, we don't have the actual type structure here yet, just the expressions, so I did not try to unpack the tuple annotation from the parameter, either. Just keeping the attribute annotation as None.

Reviewed By: sinancepel

Differential Revision: D9125576

fbshipit-source-id: ff5087924c5cb8a7fd3a10f799f4426d2a986721