Expect semicolon separated list for filter_directories
Summary:
welcomeeverythingisfine

Pyre is crashing on all single-target runs because we don't create a shared source directory for these, just defer directly to the buck-generated directory in buck-out. These sometimes have commas in the name, so when we pass a comma-separated list as `filter-directories`, everything breaks.

Deployment: Upgrade binary first, then canary and push wrapper, then deprecate old flag. We can do some more juggling in the future to get the naming back.

Reviewed By: sinancepel

Differential Revision: D9138974

fbshipit-source-id: ccea2ff74c4d1f057c28a583708243df14ab7572